dashboard pages comp setup
