--- src/app/(auth)/layout.tsx
@@ -1,8 +1,0 @@
-import React from "react"
-export default function AuthLayout({
-  children,
-}: {
-  children: React.ReactNode
-}) {
-  return <div className="min-h-screen bg-background">{children}</div>
-}
