+++ src/app/(dashboard)/settings/page.tsx
@@ -2,25 +2,16 @@
 
 import { useState } from "react";
 import { motion } from "framer-motion";
-import { useTheme } from "next-themes";
+import { AppearanceSection } from "./appearance-section";
 import {
-  Settings,
-  Bell,
-  Palette,
-  Shield,
-  Download,
-  Trash2,
-  Moon,
-  Sun,
-  Monitor,
-} from "lucide-react";
-import { Button } from "@/app/_components/ui/button";
-import { Switch } from "@/app/_components/ui/switch";
-import { Label } from "@/app/_components/ui/label";
+  NotificationsSection,
+  type NotificationSettings,
+} from "./notifications-section";
+import { PrivacySecuritySection } from "./privacy-security-section";
+import { DataManagementSection } from "./data-management-section";
 
 export default function SettingsPage() {
-  const { theme, setTheme } = useTheme();
-  const [notifications, setNotifications] = useState({
+  const [notifications, setNotifications] = useState<NotificationSettings>({
     email: true,
     push: false,
     updates: true,
@@ -28,7 +19,6 @@ export default function SettingsPage() {
 
   return (
     <div className="p-6 lg:p-8 max-w-4xl mx-auto">
-      {/* Header */}
       <motion.div
         initial={{ opacity: 0, y: -20 }}
         animate={{ opacity: 1, y: 0 }}
@@ -43,199 +33,16 @@ export default function SettingsPage() {
       </motion.div>
 
       <div className="space-y-6">
-        {/* Appearance */}
-        <motion.div
-          initial={{ opacity: 0, y: 20 }}
-          animate={{ opacity: 1, y: 0 }}
-          transition={{ delay: 0.1 }}
-          className="rounded-xl glass neon-border p-6"
-        >
-          <h2 className="font-semibold text-lg mb-4 flex items-center gap-2">
-            <Palette className="h-5 w-5 text-primary" />
-            Appearance
-          </h2>
+        <AppearanceSection />
 
-          <div className="space-y-4">
-            <div>
-              <Label className="mb-3 block">Theme</Label>
-              <div className="grid grid-cols-3 gap-3">
-                {[
-                  { value: "light", icon: Sun, label: "Light" },
-                  { value: "dark", icon: Moon, label: "Dark" },
-                  { value: "system", icon: Monitor, label: "System" },
-                ].map((option) => {
-                  const Icon = option.icon;
-                  const isSelected = theme === option.value;
+        <NotificationsSection
+          value={notifications}
+          onChange={setNotifications}
+        />
 
-                  return (
-                    <button
-                      key={option.value}
-                      onClick={() => setTheme(option.value)}
-                      className={`p-4 rounded-lg border-2 transition-all ${
-                        isSelected
-                          ? "border-primary bg-primary/10"
-                          : "border-border hover:border-primary/50 bg-muted/30"
-                      }`}
-                    >
-                      <Icon
-                        className={`h-5 w-5 mx-auto mb-2 ${
-                          isSelected ? "text-primary" : "text-muted-foreground"
-                        }`}
-                      />
-                      <span
-                        className={`text-sm font-medium ${
-                          isSelected ? "text-primary" : "text-muted-foreground"
-                        }`}
-                      >
-                        {option.label}
-                      </span>
-                    </button>
-                  );
-                })}
-              </div>
-            </div>
-          </div>
-        </motion.div>
+        <PrivacySecuritySection />
 
-        {/* Notifications */}
-        <motion.div
-          initial={{ opacity: 0, y: 20 }}
-          animate={{ opacity: 1, y: 0 }}
-          transition={{ delay: 0.2 }}
-          className="rounded-xl glass neon-border p-6"
-        >
-          <h2 className="font-semibold text-lg mb-4 flex items-center gap-2">
-            <Bell className="h-5 w-5 text-primary" />
-            Notifications
-          </h2>
-
-          <div className="space-y-4">
-            <div className="flex items-center justify-between">
-              <div>
-                <Label>Email Notifications</Label>
-                <p className="text-sm text-muted-foreground">
-                  Receive updates about your charts via email
-                </p>
-              </div>
-              <Switch
-                checked={notifications.email}
-                onCheckedChange={(checked) =>
-                  setNotifications({ ...notifications, email: checked })
-                }
-              />
-            </div>
-
-            <div className="flex items-center justify-between">
-              <div>
-                <Label>Push Notifications</Label>
-                <p className="text-sm text-muted-foreground">
-                  Get notified in your browser
-                </p>
-              </div>
-              <Switch
-                checked={notifications.push}
-                onCheckedChange={(checked) =>
-                  setNotifications({ ...notifications, push: checked })
-                }
-              />
-            </div>
-
-            <div className="flex items-center justify-between">
-              <div>
-                <Label>Product Updates</Label>
-                <p className="text-sm text-muted-foreground">
-                  News about new features and improvements
-                </p>
-              </div>
-              <Switch
-                checked={notifications.updates}
-                onCheckedChange={(checked) =>
-                  setNotifications({ ...notifications, updates: checked })
-                }
-              />
-            </div>
-          </div>
-        </motion.div>
-
-        {/* Privacy & Security */}
-        <motion.div
-          initial={{ opacity: 0, y: 20 }}
-          animate={{ opacity: 1, y: 0 }}
-          transition={{ delay: 0.3 }}
-          className="rounded-xl glass neon-border p-6"
-        >
-          <h2 className="font-semibold text-lg mb-4 flex items-center gap-2">
-            <Shield className="h-5 w-5 text-primary" />
-            Privacy & Security
-          </h2>
-
-          <div className="space-y-4">
-            <div className="flex items-center justify-between p-4 rounded-lg bg-muted/30">
-              <div>
-                <Label>Two-Factor Authentication</Label>
-                <p className="text-sm text-muted-foreground">
-                  Add an extra layer of security to your account
-                </p>
-              </div>
-              <Button variant="outline" size="sm">
-                Enable
-              </Button>
-            </div>
-
-            <div className="flex items-center justify-between p-4 rounded-lg bg-muted/30">
-              <div>
-                <Label>Active Sessions</Label>
-                <p className="text-sm text-muted-foreground">
-                  Manage devices where you are logged in
-                </p>
-              </div>
-              <Button variant="outline" size="sm">
-                View
-              </Button>
-            </div>
-          </div>
-        </motion.div>
-
-        {/* Data Management */}
-        <motion.div
-          initial={{ opacity: 0, y: 20 }}
-          animate={{ opacity: 1, y: 0 }}
-          transition={{ delay: 0.4 }}
-          className="rounded-xl glass neon-border p-6"
-        >
-          <h2 className="font-semibold text-lg mb-4 flex items-center gap-2">
-            <Settings className="h-5 w-5 text-primary" />
-            Data Management
-          </h2>
-
-          <div className="space-y-4">
-            <div className="flex items-center justify-between p-4 rounded-lg bg-muted/30">
-              <div>
-                <Label>Export Data</Label>
-                <p className="text-sm text-muted-foreground">
-                  Download all your charts and data
-                </p>
-              </div>
-              <Button variant="outline" size="sm">
-                <Download className="h-4 w-4 mr-2" />
-                Export
-              </Button>
-            </div>
-
-            <div className="flex items-center justify-between p-4 rounded-lg bg-destructive/5 border border-destructive/20">
-              <div>
-                <Label className="text-destructive">Delete Account</Label>
-                <p className="text-sm text-muted-foreground">
-                  Permanently delete your account and all data
-                </p>
-              </div>
-              <Button variant="destructive" size="sm">
-                <Trash2 className="h-4 w-4 mr-2" />
-                Delete
-              </Button>
-            </div>
-          </div>
-        </motion.div>
+        <DataManagementSection />
       </div>
     </div>
   );
